Add wanted pages list.
Fixed round-tripping of non-existent pages from editor. They were getting
broken because the editor can't handle the new_page_form links. The fix is to
just format URI-fy a non-existent page as if it does exist _for the editor
only_. The editor doesn't really need to know if a page exists or not.